third_party/WebKit/LayoutTests/fast/css/invalidation/autofill-pseudo-expected.txt

Use descendant invalidation set for :-webkit-autofill pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS [object Internals] is defined.
PASS getComputedStyle(text, '').textDecorationLine is 'none'
FAIL internals.updateStyleAndReturnAffectedElementCount() should be 1. Was 2.
PASS getComputedStyle(text, '').textDecorationLine is 'underline'
PASS successfullyParsed is true

TEST COMPLETE
